type checking

+++ chia/util/math.py
@@ -7,8 +7,8 @@ def clamp(n: int, smallest: int, largest: int) -> int:
     return max(smallest, min(n, largest))
 
 
-def make_monotonically_decreasing(seq: List[int]) -> List[int]:
-    out: List[int] = []
+def make_monotonically_decreasing(seq: List[float]) -> List[float]:
+    out: List[float] = []
     if len(seq) > 0:
         min = seq[0]
         for n in seq:
